requirements.txt
matplotlib==3.5.2
mlflow==1.26.1
numpy==1.21.6
optuna==2.10.0
pandas==1.3.5
Pillow==9.1.1
plotly==5.8.0
pytorch_lightning==1.4.7
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
gdown>=4.4.0